API

Simulation

NBodySimulator.NBodySimulation — Type

Simulation is an entity determining parameters of the experiment: time span of simulation, global physical constants, borders of the simulation cell, external magnetic or electric fields, etc. The required arguments for NBodySImulation constructor are the system to be tested and the time span of the simulation.

NBodySimulator.run_simulation — Function

Run the N-body simulation.

Bodies

NBodySimulator.Body — Type

Bodies or Particles are the objects which will interact with each other and for which the equations of Newton's 2nd law are solved during the simulation process.

Missing docstring.

Missing docstring for MassBody. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ChargedParticle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MagneticParticle. Check Documenter's build log for details.

NBodySimulator.generate_bodies_in_cell_nodes — Function

Places similar particles in the nodes of a cubic cell with their velocities distributed in accordance with the Maxwell–Boltzmann law

Potentials

NBodySimulator.PotentialParameters — Type

The potentials or force field determines the interaction of particles and, therefore, their acceleration.

Missing docstring.

Missing docstring for NBodySimulator.get_accelerating_function. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GravitationalParameters. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MagnetostaticParameters. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ElectrostaticParameters. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LennardJonesParameters. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SPCFwParameters. Check Documenter's build log for details.

Thermostats

NBodySimulator.Thermostat — Type

Usually, during the simulation, a system is required to be at a particular temperature. NBodySimulator contains several thermostats for that purpose.

NBodySimulator.NullThermostat — Type

No thermostat.

Missing docstring.

Missing docstring for AndersenThermostat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BerendsenThermostat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NoseHooverThermostat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LangevinThermostat. Check Documenter's build log for details.

Boundary Conditions

Missing docstring.

Missing docstring for CubicPeriodicBoundaryConditions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PeriodicBoundaryConditions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for InfiniteBox. Check Documenter's build log for details.

Systems

NBodySimulator.PotentialNBodySystem — Type

Structure that represents systems with a custom set of potentials. In other words, the user determines the ways in which the particles are allowed to interact. One can pass the bodies and parameters of interaction potentials into that system. In the case the potential parameters are not set, the particles will move at constant velocities without acceleration during the simulation.

Missing docstring.

Missing docstring for GravitationalSystem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for WaterSPCFw. Check Documenter's build log for details.

Analyze

NBodySimulator.SimulationResult — Type

SimulationResult sould provide an interface for working with properties of a separate particle and with physical properties of the whole system.

Missing docstring.

Missing docstring for temperature. Check Documenter's build log for details.

Missing docstring.

Missing docstring for rdf. Check Documenter's build log for details.

Missing docstring.

Missing docstring for msd. Check Documenter's build log for details.

Missing docstring.

Missing docstring for initial_energy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for kinetic_energy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for potential_energy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for total_energy. Check Documenter's build log for details.

Protein Database File

NBodySimulator.load_water_molecules_from_pdb — Function

Molecules for the SPC/Fw water model can be imported from a PDB file.

Missing docstring.

Missing docstring for save_to_pdb. Check Documenter's build log for details.